Fix annoying warning about PATH being changed in the wrong places
Apparently, the way that OMZ is loaded violates some sort of expectation about
when and where the shell PATH (and MANPATH) are updated, and this annoys
Spacemacs. The internet said that disabling the automatic PATH-getting feature
of Spacemacs' startup process, and then manually setting in my user-config, is a
valid workaround.